Distinguish exported symbols from those not exported.
The themes previously did not visually distinguish those symbols that
are exported from those that are not exported. This commit modifies
the theme to make the distinction.

The relevantFlags helper is required due to changes to
typedoc. Typedoc would previously arbitrarily narrow the array of
flags recorded in reflections to those that the default (and minimal)
themes care about. Typedoc has been changed to include all the flags
in the array and let the themes decide which flags are relevant to
them.